% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobs_colours.R
\name{Jacobs_cols}
\alias{Jacobs_cols}
\title{Function to extract Jacobs colours as hex codes}
\usage{
Jacobs_cols(...)
}
\arguments{
\item{...}{Character names of colours}
}
\value{
Colours as hex codes
}
\description{
Function to extract Jacobs colours as hex codes
}
\examples{

Jacobs_cols("Red")
Jacobs_cols("Red","darkBlue")
}
