% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobs_colours.R
\name{Jacobs_pal}
\alias{Jacobs_pal}
\title{Return function to interpolate a Jacobs color palette}
\usage{
Jacobs_pal(
  palette = "mainP1",
  reverse = FALSE,
  discrete = FALSE,
  categories = NULL,
  offset_pallete = 0,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in Jacobs_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{discrete}{Boolean indicating whether the palette is going to be used for discrete categories}

\item{categories}{Integer between 1 and 20 with number of categories if discrete is TRUE}

\item{offset_pallete}{Integer positive indicating the change of initial colour for the palette}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{

}
\description{
Return function to interpolate a Jacobs color palette
}
\examples{

}
