% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobs_colours.R
\name{scale_color_Jacobs}
\alias{scale_color_Jacobs}
\title{Colour scale constructor for Jacobs colours}
\usage{
scale_color_Jacobs(
  palette = "mainP1",
  discrete = TRUE,
  reverse = FALSE,
  categories = 5,
  offset_pallete = 0,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in Jacobs_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{offset_pallete}{}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradient(), used respectively when discrete is TRUE or FALSE}
}
\value{

}
\description{
Colour scale constructor for Jacobs colours
}
\examples{

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) +
scale_color_Jacobs(pallete)


}
