% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobs_colours.R
\name{scale_fill_Jacobs}
\alias{scale_fill_Jacobs}
\title{Fill scale constructor for Jacobs colours}
\usage{
scale_fill_Jacobs(
  palette = "mainP1",
  discrete = TRUE,
  reverse = FALSE,
  categories = 5,
  offset_pallete = 0,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in Jacobs_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\value{

}
\description{
Fill scale constructor for Jacobs colours
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4, alpha = .6) +
scale_color_Jacobs(discrete = FALSE,palette = "mainP2")
}
